move the non-Ember bits into external
